+++ hamsBE/internal/service/device_sevice.go
@@ -3,21 +3,35 @@ package service
 import (
 	"context"
 	"errors"
+	"fmt"
 	"hamstercare/internal/model"
 	"hamstercare/internal/repository"
 )
 
 type DeviceService struct {
 	DeviceRepo *repository.DeviceRepository
+	CageRepo *repository.CageRepository
 }
 
-func NewDeviceService(deviceRepo *repository.DeviceRepository) *DeviceService {
-	return &DeviceService{DeviceRepo: deviceRepo}
+func NewDeviceService(deviceRepo *repository.DeviceRepository, CageRepo *repository.CageRepository) *DeviceService {
+	return &DeviceService{DeviceRepo: deviceRepo, CageRepo: CageRepo}
 }
 
 func (s *DeviceService) CreateDevice(ctx context.Context, name, deviceType, cageID string) (*model.Device, error) {
 	if name == "" || deviceType == "" || cageID == ""{
 		return nil, errors.New("name, deviceType and cageID are required")
+	}
+
+	if err := IsValidUUID(cageID); err != nil {
+		return nil, err 
+	}
+
+	exists, err := s.CageRepo.CageExists(ctx, cageID)
+	if err != nil {
+		return nil, fmt.Errorf("error checking cage existence: %w", err)
+	}
+	if !exists {
+		return nil, fmt.Errorf("%w: cage with ID %s does not exist", ErrCageNotFound, cageID)
 	}
 
 	device, err := s.DeviceRepo.CreateDevice(ctx, name, deviceType, cageID)
@@ -56,6 +70,18 @@ func (s *DeviceService) GetDeviceByID(ctx context.Context, deviceID string) (*mo
 func (s *DeviceService) DeleteDevice(ctx context.Context, deviceID string) error {
 	if deviceID == "" {
 		return errors.New("deviceID is required")
+	}
+
+	if err := IsValidUUID(deviceID); err != nil {
+		return err 
+	}
+	
+	exists, err := s.DeviceRepo.DeviceExists(ctx, deviceID)
+	if err != nil {
+		return fmt.Errorf("error checking device existence: %w", err)
+	}
+	if !exists {
+		return fmt.Errorf("%w: device with ID %s does not exist", ErrDeviceNotFound, deviceID)
 	}
 
 	return s.DeviceRepo.DeleteDeviceByID(ctx, deviceID)
